requirements.txt
streamlit==1.28.0
pillow==10.0.1
pytesseract==0.3.11
pdf2image==1.16.3
pandas==2.1.1
opencv-python==4.8.1.78
numpy==1.27.4
rapidfuzz==2.16.0
